extract dsl info from speedport engineer menu
First version
This version supports authentication and can parse the engineer dsl
info.

Authentication works by fetching the login.json which contains the
challenge. This challenge then gets sha256 summed, together with the
password.

With the hashed_pw as a POST parameter the login.json returns a
successful login and a session cookie.
This cookie can the be added to any further request to authenticate.